Update code check the sonar report v1.3

+++ product/src/main/java/com/yas/product/controller/ProductAttributeValueController.java
@@ -19,6 +19,7 @@ import org.springframework.web.util.UriComponentsBuilder;
 
 import javax.validation.Valid;
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 public class ProductAttributeValueController {
@@ -93,9 +94,9 @@ public class ProductAttributeValueController {
             @ApiResponse(responseCode = "404", description = "Not found", content = @Content(schema = @Schema(implementation = ErrorVm.class))),
             @ApiResponse(responseCode = "400", description = "Bad request", content = @Content(schema = @Schema(implementation = ErrorVm.class)))})
     public ResponseEntity<Void> deleteProductAttributeValueById(@PathVariable Long id) {
-        productAttributeValueRepository
-                .findById(id)
-                .orElseThrow(()-> new NotFoundException(String.format("Product attribute value %s is not found", id)));
+        Optional<ProductAttributeValue> productAttributeValue = productAttributeValueRepository.findById(id);
+        if(productAttributeValue.isEmpty())
+            throw new  NotFoundException(String.format("Product attribute value %s is not found", id));
         productAttributeValueRepository.deleteById(id);
         return ResponseEntity.noContent().build();
     }
